v1-skeleton/requirements.txt
# AI 知识库项目依赖
requests>=2.31.0
python-dotenv>=1.0.0
tavily-python>=0.5.0

# 可选：用于优化性能
# aiohttp>=3.9.0
# asyncio>=3.4.3